+++ packages/web3-errors/CHANGELOG.md
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   The abstract class `Web3Error` is renamed to `BaseWeb3Error` (#5771)
 -   Renamed TransactionRevertError to TransactionRevertInstructionError to remain consistent with 1.x
+-   Using `MaxAttemptsReachedOnReconnectingError` with the same message for 1.x but also adding the `maxAttempts` (#5894)
 
 ### Added